Move peer to bin/glusterd2/peer
