+++ bin/glusterd2/commands/peers/deletepeer.go
@@ -4,9 +4,9 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/peer"
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/bin/glusterd2/store"
-	"github.com/gluster/glusterd2/peer"
 	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/volume"
