+++ bin/glusterd2/peer/self.go
