+++ volume/volume_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/gluster/glusterd2/peer"
+	"github.com/gluster/glusterd2/bin/glusterd2/peer"
 	"github.com/gluster/glusterd2/pkg/errors"
 	"github.com/gluster/glusterd2/tests"
